refactor(website): cleaned up last legacy components
