--- aksel.nav.no/website/components/website-modules/StatusTag.tsx
@@ -1,47 +1,0 @@
-import { Tag } from "@navikt/ds-react";
-
-export const StatusTag = ({
-  status,
-  showStable = false,
-  size = "small",
-  ...rest
-}: {
-  size?: "small" | "xsmall";
-  status: string;
-  showStable?: boolean;
-}) => {
-  switch (status) {
-    case "preview":
-      return (
-        <Tag size={size} data-color="meta-purple" {...rest}>
-          Preview
-        </Tag>
-      );
-    case "beta":
-      return (
-        <Tag size={size} data-color="meta-purple" {...rest}>
-          Beta
-        </Tag>
-      );
-    case "new":
-      return (
-        <Tag data-color="info" size={size} {...rest}>
-          Ny
-        </Tag>
-      );
-    case "ready":
-      return showStable ? (
-        <Tag data-color="success" size={size} {...rest}>
-          Stabil
-        </Tag>
-      ) : null;
-    case "deprecated":
-      return (
-        <Tag data-color="neutral" size={size} {...rest}>
-          Avviklet
-        </Tag>
-      );
-    default:
-      return null;
-  }
-};
